Standalone engine and code generation updates for processing validation and result logging.
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@444 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 8c64db7645519bacb9fcb5ce0c2420e33ca4ce11

+++ mmdb/dbtoaster/prototype/compiler/alpha3/examples/vwap.sql
@@ -1,4 +1,4 @@
-CREATE TABLE BIDS (t int, id int, broker_id int, p int, v int)
+CREATE TABLE BIDS (t double, id int, broker_id int, p double, v double)
     FROM 'file'
     SOURCE 'DBToaster::DemoDatasets::OrderbookFileStream'
     ARGS '"/Users/yanif/tmp/data/cleanedData.csv",10000'
